Fix Shop Panel

+++ Assets/Scripts/SO/ItemSO.cs
@@ -8,6 +8,6 @@ public class ItemSO : ScriptableObject
     public string nameItem;
     public bool isStackable;
     public Sprite icon;
-    public float buyPrice;
-    public float sellPrice;
+    public int buyPrice;
+    public int sellPrice;
 }
